add a title to map and scale to legend
